requirements.txt
fastapi==0.109.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
uvicorn==0.24.0.post1
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
argon2==0.1.10
python-multipart==0.0.6
beanie==1.23.1
python-decouple==3.6
pydantic_settings==2.0.3
pydantic[email]==2.4.2
aiomysql==0.2.0
aiosqlite==0.19.0
boto3==1.28.84
asyncpg==0.29.0
sqlmodel==0.0.18
pyjwt==2.8.0
email-validator==2.1.1
annotated-types==0.7.0